fix(dagre): Fixed the error positions of nodes with dagre layout.
The G6/ dagre layout uses the original index of nodes to set the {x, y}.

But in dagre.js, the g.nodes() and g.edges() return arrays not the same order with out original nodes and edges.

So we should use the node.id and edge.source+edge.target to get the original model to set their positions.

Fixed #1032